src/test/java/com/example/service/ProductServiceTest.java
package com.example.service;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.assertNotNull;


@RunWith(SpringRunner.class)
@SpringBootTest
public class ProductServiceTest {

    @Autowired
    private ProductService productService;

    @Test
    public void getUsersFromApi() {
        assertNotNull(productService.getProductsPathFromApi());
    }
    @Test
    public void saveToExcel() {
        productService.saveToExcel();
    }

}